Clippy for Rust 1.66 (#1284)

+++ lib/collection/src/operations/conversions.rs
@@ -166,7 +166,7 @@ impl From<CollectionInfo> for api::grpc::qdrant::CollectionInfo {
                     max_segment_size: config.optimizer_config.max_segment_size.map(|x| x as u64),
                     memmap_threshold: config.optimizer_config.memmap_threshold.map(|x| x as u64),
                     indexing_threshold: Some(config.optimizer_config.indexing_threshold as u64),
-                    flush_interval_sec: Some(config.optimizer_config.flush_interval_sec as u64),
+                    flush_interval_sec: Some(config.optimizer_config.flush_interval_sec),
                     max_optimization_threads: Some(
                         config.optimizer_config.max_optimization_threads as u64,
                     ),
